Upgrade script to python3 and remove unused import
Ubuntu 14.04 LTS defaults to python3. This required
updating both my_init and setuser. Also, my_init
referenced the deprecated posixfile module, which
was removed in python3. It was not used anyway, so
I simply removed the import.

For the curious, posixfile was deprecated by
fcntl.lockf, which does the same thing, but
more
portably.

https://docs.python.org/2/library/fcntl.html#fcntl.lockf

Signed-off-by: Christian Romney <cromney@pointslope.com>