Fix README changed on PR-182
There no point in changing the build process of an image based on this
baseimage, since the services would be already installed (or note) on
the original image. The point for optional services installation is
just useful when build your own image based on this baseimage source
repository, change the `image/buildconfig` file.